cmPackageInfoReader: Fix IMPORTED_CONFIGURATIONS
Rework how we assign imported configurations to only add configurations
that are actually imported. This requires a certain amount of cleverness
to keep the order consistent with the package's specified default
configurations, but doing this is important now that configuration
selection (see policies CMP0199 and CMP0200) is more reliant on the
IMPORTED_CONFIGURATIONS property being accurate, rather than focusing on
whether configuration-specific properties are set.