FindwxWidgets: updates for wxWidgets 3.3
wxWidgets 3.3 changed some win32 libraries from dynamic to static loading.
Update the list of win32 libraries to the list used upstream.
wsock32 is not used in wxWidgets 3.3 anymore, but keep it for older wxWidgets versions.

wxWidgets 3.3 also adds the wxwebp library. Add it, and its sub-libraries, to wxWidgets_COMMON_LIBRARIES.