GoogleTest: Update policies to CMake 4.2
In fc7aa3cd69 (tests: Preserve empty arguments in test command lines,
2024-09-30), the CMP0174 policy setting had to be explicitly handled in
GoogleTestAddTests.cmake because it was part of the same release.
Now that we're past that release, we can remove that and update the
policies based on the CMake version instead. This removes the TODO
added in that original commit.

We also update the policy settings in GoogleTest.cmake so that we are
using consistent policy settings across the GoogleTest implementation.