cmake --build: Factor out a struct holding build parameters
This commit extracts the set of parameters required to build a CMake
project into a dedicated class.  While this class is currently just a
wrapper for the data, the goal is to eventually implement a validate()
method which verifies that a given configuration contains the minimum
set of information required to execute a build, and that any optional
options are compatible with one another.