CPS: Be less optimistic about what we can read
Previously, `cmPackageInfoReader` made the wildly optimistic assumption
that it will know how to read any CPS version that is less than 1.0
(note, the current version as of writing is 0.13). This is likely not a
safe assumption. Assume instead that we only know how to read the
current minor version, i.e. 0.13.x. This still allows room for the
specification to make non-breaking changes, while also providing a
better path for making breaking changes without CMake wrongly thinking
it still understands what it's reading.